implement autogen tables for fluid_rvoice_dsp
