+++ src/rvoice/fluid_phase.h
@@ -22,8 +22,6 @@
 #ifndef _FLUID_PHASE_H
 #define _FLUID_PHASE_H
 
-#include "config.h"
-
 /*
  *  phase
  */
@@ -31,7 +29,7 @@
 #define FLUID_INTERP_BITS        8
 #define FLUID_INTERP_BITS_MASK   0xff000000
 #define FLUID_INTERP_BITS_SHIFT  24
-#define FLUID_INTERP_MAX         256
+
 
 #define FLUID_FRACT_MAX ((double)4294967296.0)
