+++ src/rvoice/fluid_rvoice_dsp_tables.h
@@ -1,0 +1,8 @@
+
+#ifndef _FLUID_RVOICE_DSP_TABLES_H
+#define _FLUID_RVOICE_DSP_TABLES_H
+
+#define FLUID_INTERP_MAX         256
+#define SINC_INTERP_ORDER 7	/* 7th order constant */
+
+#endif
